@@ -1,5 +1,5 @@
-import { Avatar } from '@material-ui/core'
-import { InsertEmoticon, PhotoLibrary, Videocam } from '@material-ui/icons';
+import { Avatar } from '@mui/material'
+import { InsertEmoticon, PhotoLibrary, Videocam } from'@mui/icons-material'
 import React, { useState } from 'react'
 import './Question.css'
 // import {useStateValue} from '../StateProvider'
@@ -10,7 +10,7 @@ const Question = () => {
 
   const [input,setInput] = useState('');
   const [image,setImage] = useState(null);
-//   const [{user},dispatch] = useStateValue()
+  //const [{user},dispatch] = useStateValue()
 
   const handleChange = (e) => {
       if(e.target.files[0]) {
@@ -18,65 +18,66 @@ const Question = () => {
       }
   }
 
-//   const handleSubmit = async (e) => {
-//       e.preventDefault();
-//       console.log("Submitting");
+  const handleSubmit = async (e) => {
+      e.preventDefault();
+      console.log("Submitting");
 
-//       if(image) {
-//         const imgForm = new FormData();
-//         imgForm.append('file',image,image.name)
+    //   if(image) {
+    //     const imgForm = new FormData();
+    //     imgForm.append('file',image,image.name)
 
-//         axios.post('/upload/image',imgForm, {
-//           headers:{
-//             'accept': 'application/json',
-//             'Accept-Language':'en-US,en;q=0.8',
-//             'Content-Type': `multipart/form-data; boundary=${imgForm._boundary}`
-//           }
-//         }).then((res)=>{
-//             console.log(res.data)
+    //     axios.post('/upload/image',imgForm, {
+    //       headers:{
+    //         'accept': 'application/json',
+    //         'Accept-Language':'en-US,en;q=0.8',
+    //         'Content-Type': `multipart/form-data; boundary=${imgForm._boundary}`
+    //       }
+    //     }).then((res)=>{
+    //         console.log(res.data)
 
-//             const postData = {
-//               text:input,
-//               imgName:res.data.filename,
-//               user:user.displayName,
-//               avatar:user.photoURL,
-//               timestamp: Date.now()
-//             }
-//             console.log(postData);
-//             savePost(postData)
+    //         const postData = {
+    //           text:input,
+    //           imgName:res.data.filename,
+    //           user:user.displayName,
+    //           avatar:user.photoURL,
+    //           timestamp: Date.now()
+    //         }
+    //         console.log(postData);
+    //         savePost(postData)
 
-//         })
-//       }
-//       else
-//       {
-//           const postData = {
-//             text:input,
-//             user:user.displayName,
-//             avatar:user.photoURL,
-//             timestamp: Date.now()
-//           }
-//           console.log(postData);
-//           savePost(postData)
+    //     })
+    //   }
+    //   else
+    //   {
+    //       const postData = {
+    //         text:input,
+    //         user:user.displayName,
+    //         avatar:user.photoURL,
+    //         timestamp: Date.now()
+    //       }
+    //       console.log(postData);
+    //       savePost(postData)
 
-//       }
+    //   }
 
-//       setImage('');
-//       setInput('');
+      setImage('');
+      setInput('');
       
-//   }
+  }
 
-//   const savePost = async (postData) =>{
-//     await axios.post('/upload/post',postData)
-//           .then((resp)=>{
-//             console.log(resp);
-//           })
-//   }
+  const savePost = async (postData) =>{
+    console.log(postData)
+    // await axios.post('/upload/post',postData)
+    //       .then((resp)=>{
+    //         console.log(resp);
+    //       })
+  }
 
 
   return (
     <div className='question'>
       <div className='question__top'>
-        {/* <Avatar src={user.photoURL}/> */}
+        <Avatar src={"user.photoURL"}/>
 
         <form>
           <input type='text' placeholder='Submit Query here ...'
@@ -88,7 +89,7 @@ const Question = () => {
           <input type='file' 
            className='question__file' onChange={handleChange}/>
 
-          <button onClick={"handleSubmit"} type='submit'>Upload Crop Details</button>
+          <button onClick={handleSubmit} type='submit'>Upload Crop Details</button>
         </form>
 
       </div>
